+++ src/components/pages/user/DeleteUser.jsx
@@ -13,8 +13,10 @@ const DeleteUser = () => {
     const params = useParams();
     useEffect(()=> {
         if(verifietoken()){
+            if (window.confirm('Vous voulez vraiment supprimer l`utilisateur?')) {
             deleteUser(params.id)
             alert("utilisateur est bien supprimé")
+            }
             history.push("/user-list");
         }
         else{
